requirements.txt
google-cloud-bigquery>=3.11.0
gspread>=5.10.0
google-auth>=2.22.0
